+++ CHANGELOG.md
@@ -8,7 +8,8 @@ All notable changes to this project will be documented in this file.
 - download verification by checking if downloaded files have the expected size
 - decompress compressed g-band and z-band files to speed up subsequent file loading times
 - post-download cutout creation, store in hdf5 files
-- added retry logic for downloads
+- retry logic for downloads
+- plotting routine for cutouts (unionsdata plot)
 
 ## [0.1.1] - 2025-10-31